refactored logic for starting pet profile from community
