+++ client/android/app/src/main/java/com/sju18001/petmanagement/ui/community/post/PostFragment.kt
@@ -32,6 +32,7 @@ import com.sju18001.petmanagement.restapi.SessionManager
 import com.sju18001.petmanagement.restapi.dao.Pet
 import com.sju18001.petmanagement.restapi.dao.Post
 import com.sju18001.petmanagement.restapi.dto.*
+import com.sju18001.petmanagement.ui.community.CommunityUtil
 import com.sju18001.petmanagement.ui.community.CommunityViewModel
 import com.sju18001.petmanagement.ui.community.comment.CommunityCommentActivity
 import com.sju18001.petmanagement.ui.community.post.createUpdatePost.CreateUpdatePostActivity
@@ -317,14 +318,14 @@ class PostFragment : Fragment() {
             override fun fetchPetPhotoAndStartPetProfileFragment(holder: PostListAdapter.ViewHolder, pet: Pet) {
                 // 사진이 없을 때는 fetch 없이 프래그먼트 시작
                 if(pet.photoUrl == null){
-                    startPetProfileFragment(holder, pet, null)
+                    CommunityUtil.startPetProfileFragmentFromCommunity(holder.itemView.context, pet, null)
                     return
                 }
 
                 val call = RetrofitBuilder.getServerApiWithToken(SessionManager.fetchUserToken(requireContext())!!)
                     .fetchPetPhotoReq(FetchPetPhotoReqDto(pet.id))
                 ServerUtil.enqueueApiCall(call, isViewDestroyed, requireContext(), { response ->
-                    startPetProfileFragment(holder, pet, response.body()!!.bytes())
+                    CommunityUtil.startPetProfileFragmentFromCommunity(holder.itemView.context, pet, response.body()!!.bytes())
                 }, {}, {})
             }
         }
@@ -484,41 +485,6 @@ class PostFragment : Fragment() {
             }
         })
             .create().show()
-    }
-
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun startPetProfileFragment(holder: PostListAdapter.ViewHolder, pet: Pet, photoByteArray: ByteArray?){
-        // set pet values to Intent
-        val petProfileIntent = Intent(holder.itemView.context, MyPetActivity::class.java)
-        if(photoByteArray != null) {
-            Util.saveByteArrayToSharedPreferences(requireContext(), requireContext().getString(R.string.pref_name_byte_arrays),
-                requireContext().getString(R.string.data_name_community_selected_pet_photo), photoByteArray)
-        }
-        else {
-            Util.saveByteArrayToSharedPreferences(requireContext(), requireContext().getString(R.string.pref_name_byte_arrays),
-                requireContext().getString(R.string.data_name_community_selected_pet_photo), null)
-        }
-        petProfileIntent.putExtra("petId", pet.id)
-        petProfileIntent.putExtra("petName", pet.name)
-        petProfileIntent.putExtra("petBirth", pet.birth)
-        petProfileIntent.putExtra("petSpecies", pet.species)
-        petProfileIntent.putExtra("petBreed", pet.breed)
-        val petGender = if(pet.gender) {
-            holder.itemView.context.getString(R.string.pet_gender_female_symbol)
-        }
-        else {
-            holder.itemView.context.getString(R.string.pet_gender_male_symbol)
-        }
-        val petAge = Period.between(LocalDate.parse(pet.birth), LocalDate.now()).years.toString()
-        petProfileIntent.putExtra("petGender", petGender)
-        petProfileIntent.putExtra("petAge", petAge)
-        petProfileIntent.putExtra("petMessage", pet.message)
-
-        // open activity
-        petProfileIntent.putExtra("fragmentType", "pet_profile_community")
-        holder.itemView.context.startActivity(petProfileIntent)
-        (requireContext() as Activity).overridePendingTransition(R.anim.enter_from_right, R.anim.exit_to_left)
     }
